requirements.txt
ffmpeg-python==0.2.0
celery>=5.3
redis>=5.0
ray[default]>=2.10
prometheus-client>=0.20


